requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
supabase==2.3.4
python-dateutil==2.8.2
pydantic==2.5.0
gunicorn==21.2.0
